+++ dispute.go
@@ -60,6 +60,14 @@ type Dispute struct {
 	UpdatedAt         *time.Time                   `xml:"updated-at"`
 }
 
+type DisputeStatusHistoryEvent struct {
+	XMLName          string     `xml:"status-history"`
+	DisbursementDate string     `xml:"disbursement-date"`
+	EffectiveDate    string     `xml:"effective-date"`
+	Status           string     `xml:"status"`
+	Timestamp        *time.Time `xml:"timestamp"`
+}
+
 type DisputeTransaction struct {
 	XMLName                  string     `xml:"transaction"`
 	Amount                   *Decimal   `xml:"amount"`
